refactor: Move ProjectsAdapter to seperate file
